@@ -32,7 +32,7 @@
     },
     "prefer-stable": true,
     "require": {
-        "php": "^7.2",
+        "php": "^7.4",
         "ext-curl": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
