@@ -18,25 +18,16 @@ use Zoho\Desk\Model\DataObjectInterface;
 
 final class CreateOperation implements CreateOperationInterface
 {
-    /**
-     * @var RequestBuilder
-     */
-    private $requestBuilder;
+    private RequestBuilder $requestBuilder;
 
-    /**
-     * @var DataObjectFactory
-     */
-    private $dataObjectFactory;
+    private DataObjectFactory $dataObjectFactory;
 
-    /**
-     * @var string
-     */
-    private $entityType;
+    private string $entityType;
 
     /**
      * @var string[]
      */
-    private $arguments;
+    private array $arguments;
 
     public function __construct(
         RequestBuilder $requestBuilder,
