@@ -19,25 +19,16 @@ use function array_merge;
 
 final class ReadOperation implements ReadOperationInterface
 {
-    /**
-     * @var RequestBuilder
-     */
-    private $requestBuilder;
+    private RequestBuilder $requestBuilder;
 
-    /**
-     * @var DataObjectFactory
-     */
-    private $dataObjectFactory;
+    private DataObjectFactory $dataObjectFactory;
 
-    /**
-     * @var string
-     */
-    private $entityType;
+    private string $entityType;
 
     /**
      * @var string[]
      */
-    private $arguments;
+    private array $arguments;
 
     public function __construct(
         RequestBuilder $requestBuilder,
